+++ testcrate/tests/test.rs
@@ -81,6 +81,23 @@ fn to_hex_string() {
 }
 
 #[test]
+fn format() {
+    let x = u256!(0x0123456789abcedffdecba98765432100f1e2d3c4b5a697813375ca1ab1e1337);
+    assert_eq!(
+        format!("{}", x),
+        "514631507721406819000807973055347587171986281168858714843337847970547503927"
+    );
+    assert_eq!(
+        format!("{:x}", x),
+        "123456789abcedffdecba98765432100f1e2d3c4b5a697813375ca1ab1e1337"
+    );
+    assert_eq!(
+        format!("{:#x}", x),
+        "0x123456789abcedffdecba98765432100f1e2d3c4b5a697813375ca1ab1e1337"
+    );
+}
+
+#[test]
 #[should_panic]
 fn len_0() {
     let _ = Uint::<0>::from_u64_array([]);