+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.1] - 2022-2-28
+### Fixes
+- Fixed formatting impls to use `pad_integral`. The non-alternate form of `LowerHex` (the one
+  involving the "{:x}" format string) now does not include the leading "0x". Only the alternate
+  "{:#x}" format should include the leading "0x".
+
 ## [0.2.0] - 2022-2-25
 ### Fixes
 - Casting is now correct on big endian systems